Moved default error messages code to /imports/startup/client/i18n.js.
